+++ DroneDelivery/App/GamePhase.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public enum GamePhase: Sendable {
+    case menu, briefing, preflight, flying, delivering, returning, debrief
+}